sccs: correct error handling for mktime()
The mktime() function is not guaranteed to leave errno
untouched when it is succesful and neither does a -1
return value correctly indicate that the function failed.

This mishandling was exposed in a recent tzcode update on
FreeBSD, which caused mktime() to spuriously set errno for
the date corresponding to a UNIX time of -1.

The correct way to tell if the function succeeded is to
check if it wrote the week and year day fields back to the
tm structure.

Fixes		#66