Ensure max VMG size
Some DVD show the following error: 
mkisofs: Implementation botch. Video pad for file VIDEO_TS.BUP is -1
mkisofs: Either the *.IFO file is bad or you found a mkisofs bug.

Looking at the information in the IFO, it appears that vmg_last_sector is equal to the value of the first VTS sector when it should be less than one.

Using a variable max_size_vmg, it becomes easy to make sure the size of VMG, encompassing VIDEO_TS.IFO, VIDEO_TS.VOB and VIDEO_TS.BUP, stays below the first sector of the VTSs.